Make VcapServicesParser use CFEnvParser
